On the 'move-tracking-2' branch: remove a redundant cast.
* subversion/libsvn_delta/branch.c
  (svn_branch_repos_get_revision): Remove a cast.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/move-tracking-2@1688214 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_delta/branch.c
@@ -69,7 +69,7 @@ svn_branch_repos_get_revision(const svn_branch_repos_t *repos,
                               svn_revnum_t revnum)
 {
   assert(revnum < repos->rev_roots->nelts);
-  return svn_array_get(repos->rev_roots, (int)revnum);
+  return svn_array_get(repos->rev_roots, revnum);
 }
 
 struct svn_branch_state_t *
